requirements.txt
openai
python-dotenv
pandas
streamlit
pydantic
chromadb
pypdf2